refactor(edges): Change base edge to functional component
# What's changed?

* Change base edge into a functional component (doesn't need a state)
* Export base edge for convenience
